@@ -7,15 +7,9 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.media.RatingCompat;
 import android.support.v4.media.session.PlaybackStateCompat;
-
 import com.facebook.react.bridge.Promise;
-import com.facebook.react.modules.network.OkHttpClientProvider;
 import com.facebook.react.views.imagehelper.ResourceDrawableIdHelper;
-import com.google.android.exoplayer2.ext.okhttp.OkHttpDataSourceFactory;
-import com.google.android.exoplayer2.upstream.DataSource;
-import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
-import com.google.android.exoplayer2.upstream.HttpDataSource;
+import com.google.android.exoplayer2.upstream.RawResourceDataSource;
 
 /**
  * @author Guichaguri
@@ -24,19 +18,15 @@ public class Utils {
 
     public static final String EVENT_INTENT = "com.guichaguri.trackplayer.event";
     public static final String CONNECT_INTENT = "com.guichaguri.trackplayer.connect";
+    public static final String NOTIFICATION_CHANNEL = "com.guichaguri.trackplayer";
     public static final String LOG = "RNTrackPlayer";
 
-    public static Runnable toRunnable(final Promise promise) {
-        return new Runnable() {
-            @Override
-            public void run() {
-                promise.resolve(null);
-            }
-        };
+    public static Runnable toRunnable(Promise promise) {
+        return () -> promise.resolve(null);
     }
 
     public static long toMillis(double seconds) {
-        return (long) (seconds * 1000);
+        return (long)(seconds * 1000);
     }
 
     public static double toSeconds(long millis) {
@@ -44,7 +34,7 @@ public class Utils {
     }
 
     public static boolean isLocal(Uri uri) {
-        if (uri == null) return false;
+        if(uri == null) return false;
 
         String scheme = uri.getScheme();
         String host = uri.getHost();
@@ -53,6 +43,7 @@ public class Utils {
                 scheme.equals(ContentResolver.SCHEME_FILE) ||
                 scheme.equals(ContentResolver.SCHEME_ANDROID_RESOURCE) ||
                 scheme.equals(ContentResolver.SCHEME_CONTENT) ||
+                scheme.equals(RawResourceDataSource.RAW_RESOURCE_SCHEME) ||
                 scheme.equals("res") ||
                 host == null ||
                 host.equals("localhost") ||
@@ -61,23 +52,26 @@ public class Utils {
     }
 
     public static Uri getUri(Context context, Bundle data, String key) {
-        if (!data.containsKey(key)) return null;
+        if(!data.containsKey(key)) return null;
         Object obj = data.get(key);
 
-        if (obj instanceof String) {
+        if(obj instanceof String) {
             // Remote or Local Uri
 
-            return Uri.parse((String) obj);
+            if(((String)obj).trim().isEmpty())
+                throw new RuntimeException("The URL cannot be empty");
 
-        } else if (obj instanceof Bundle) {
+            return Uri.parse((String)obj);
+
+        } else if(obj instanceof Bundle) {
             // require/import
 
-            String uri = ((Bundle) obj).getString("uri");
+            String uri = ((Bundle)obj).getString("uri");
 
             ResourceDrawableIdHelper helper = ResourceDrawableIdHelper.getInstance();
             int id = helper.getResourceDrawableId(context, uri);
 
-            if (id > 0) {
+            if(id > 0) {
                 // In production, we can obtain the resource uri
                 Resources res = context.getResources();
 
@@ -97,6 +91,23 @@ public class Utils {
         return null;
     }
 
+    public static int getRawResourceId(Context context, Bundle data, String key) {
+        if(!data.containsKey(key)) return 0;
+        Object obj = data.get(key);
+
+        if(!(obj instanceof Bundle)) return 0;
+        String name = ((Bundle)obj).getString("uri");
+
+        if(name == null || name.isEmpty()) return 0;
+        name = name.toLowerCase().replace("-", "_");
+
+        try {
+            return Integer.parseInt(name);
+        } catch (NumberFormatException ex) {
+            return context.getResources().getIdentifier(name, "raw", context.getPackageName());
+        }
+    }
+
     public static boolean isPlaying(int state) {
         return state == PlaybackStateCompat.STATE_PLAYING || state == PlaybackStateCompat.STATE_BUFFERING;
     }
@@ -110,13 +121,13 @@ public class Utils {
     }
 
     public static RatingCompat getRating(Bundle data, String key, int ratingType) {
-        if (!data.containsKey(key)) {
+        if(!data.containsKey(key) || ratingType == RatingCompat.RATING_NONE) {
             return RatingCompat.newUnratedRating(ratingType);
-        } else if (ratingType == RatingCompat.RATING_HEART) {
+        } else if(ratingType == RatingCompat.RATING_HEART) {
             return RatingCompat.newHeartRating(data.getBoolean(key, true));
-        } else if (ratingType == RatingCompat.RATING_THUMB_UP_DOWN) {
+        } else if(ratingType == RatingCompat.RATING_THUMB_UP_DOWN) {
             return RatingCompat.newThumbRating(data.getBoolean(key, true));
-        } else if (ratingType == RatingCompat.RATING_PERCENTAGE) {
+        } else if(ratingType == RatingCompat.RATING_PERCENTAGE) {
             return RatingCompat.newPercentageRating(data.getFloat(key, 0));
         } else {
             return RatingCompat.newStarRating(ratingType, data.getFloat(key, 0));
@@ -124,28 +135,18 @@ public class Utils {
     }
 
     public static void setRating(Bundle data, String key, RatingCompat rating) {
-        if (!rating.isRated()) return;
+        if(!rating.isRated()) return;
         int ratingType = rating.getRatingStyle();
 
-        if (ratingType == RatingCompat.RATING_HEART) {
+        if(ratingType == RatingCompat.RATING_HEART) {
             data.putBoolean(key, rating.hasHeart());
-        } else if (ratingType == RatingCompat.RATING_THUMB_UP_DOWN) {
+        } else if(ratingType == RatingCompat.RATING_THUMB_UP_DOWN) {
             data.putBoolean(key, rating.isThumbUp());
-        } else if (ratingType == RatingCompat.RATING_PERCENTAGE) {
+        } else if(ratingType == RatingCompat.RATING_PERCENTAGE) {
             data.putDouble(key, rating.getPercentRating());
         } else {
             data.putDouble(key, rating.getStarRating());
         }
-    }
-
-    public static DataSource.Factory buildDataSourceFactory(Context context, DefaultBandwidthMeter bandwidthMeter) {
-        Context appContext = context.getApplicationContext();
-        return new DefaultDataSourceFactory(appContext, bandwidthMeter,
-                buildHttpDataSourceFactory(appContext, bandwidthMeter));
-    }
-
-    private static HttpDataSource.Factory buildHttpDataSourceFactory(Context context, DefaultBandwidthMeter bandwidthMeter) {
-        return new OkHttpDataSourceFactory(OkHttpClientProvider.getOkHttpClient(), "bookln_saas", bandwidthMeter);
     }
 
 }
