@@ -1,7 +1,10 @@
 package com.guichaguri.trackplayer.service;
 
 import android.annotation.SuppressLint;
+import android.content.BroadcastReceiver;
 import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
 import android.media.AudioAttributes;
 import android.media.AudioFocusRequest;
 import android.media.AudioManager;
@@ -14,7 +17,6 @@ import android.os.PowerManager;
 import android.os.PowerManager.WakeLock;
 import android.support.annotation.RequiresApi;
 import android.util.Log;
-
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.DefaultLoadControl;
 import com.google.android.exoplayer2.DefaultRenderersFactory;
@@ -26,6 +28,7 @@ import com.guichaguri.trackplayer.module.MusicEvents;
 import com.guichaguri.trackplayer.service.metadata.MetadataManager;
 import com.guichaguri.trackplayer.service.models.Track;
 import com.guichaguri.trackplayer.service.player.ExoPlayback;
+import com.guichaguri.trackplayer.service.player.LocalPlayback;
 
 import static com.google.android.exoplayer2.DefaultLoadControl.*;
 
@@ -39,24 +42,34 @@ public class MusicManager implements OnAudioFocusChangeListener {
     private final WakeLock wakeLock;
     private final WifiLock wifiLock;
 
-//    private MetadataManager metadata;
+    private MetadataManager metadata;
     private ExoPlayback playback;
 
     @RequiresApi(26)
     private AudioFocusRequest focus = null;
     private boolean hasAudioFocus = false;
 
+    private BroadcastReceiver noisyReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            service.emit(MusicEvents.BUTTON_PAUSE, null);
+        }
+    };
+    private boolean receivingNoisyEvents = false;
+
+    private boolean stopWithApp = false;
+
     @SuppressLint("InvalidWakeLockTag")
     public MusicManager(MusicService service) {
         this.service = service;
-//        this.metadata = new MetadataManager(service, this);
+        this.metadata = new MetadataManager(service, this);
 
-        PowerManager powerManager = (PowerManager) service.getSystemService(Context.POWER_SERVICE);
+        PowerManager powerManager = (PowerManager)service.getSystemService(Context.POWER_SERVICE);
         wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "track-player-wake-lock");
         wakeLock.setReferenceCounted(false);
 
         // Android 7: Use the application context here to prevent any memory leaks
-        WifiManager wifiManager = (WifiManager) service.getApplicationContext().getSystemService(Context.WIFI_SERVICE);
+        WifiManager wifiManager = (WifiManager)service.getApplicationContext().getSystemService(Context.WIFI_SERVICE);
         wifiLock = wifiManager.createWifiLock(WifiManager.WIFI_MODE_FULL, "track-player-wifi-lock");
         wifiLock.setReferenceCounted(false);
     }
@@ -65,82 +78,104 @@ public class MusicManager implements OnAudioFocusChangeListener {
         return playback;
     }
 
+    public boolean shouldStopWithApp() {
+        return stopWithApp;
+    }
+
+    public void setStopWithApp(boolean stopWithApp) {
+        this.stopWithApp = stopWithApp;
+    }
+
     public MetadataManager getMetadata() {
-        return null;
+        return metadata;
     }
 
     public void switchPlayback(ExoPlayback playback) {
-        if (this.playback != null) {
+        if(this.playback != null) {
+            this.playback.stop();
             this.playback.destroy();
         }
 
         this.playback = playback;
+
+        if(this.playback != null) {
+            this.playback.initialize();
+        }
     }
 
-    public ExoPlayback createLocalPlayback(Bundle options) {
-        int minBuffer = (int) Utils.toMillis(options.getDouble("minBuffer", Utils.toSeconds(DEFAULT_MIN_BUFFER_MS)));
-        int maxBuffer = (int) Utils.toMillis(options.getDouble("maxBuffer", Utils.toSeconds(DEFAULT_MAX_BUFFER_MS)));
-        int playBuffer = (int) Utils.toMillis(options.getDouble("playBuffer", Utils.toSeconds(DEFAULT_BUFFER_FOR_PLAYBACK_MS)));
-        long cacheMaxSize = (long) (options.getDouble("maxCacheSize", 0) * 1024);
+    public LocalPlayback createLocalPlayback(Bundle options) {
+        int minBuffer = (int)Utils.toMillis(options.getDouble("minBuffer", Utils.toSeconds(DEFAULT_MIN_BUFFER_MS)));
+        int maxBuffer = (int)Utils.toMillis(options.getDouble("maxBuffer", Utils.toSeconds(DEFAULT_MAX_BUFFER_MS)));
+        int playBuffer = (int)Utils.toMillis(options.getDouble("playBuffer", Utils.toSeconds(DEFAULT_BUFFER_FOR_PLAYBACK_MS)));
+        int backBuffer = (int)Utils.toMillis(options.getDouble("backBuffer", Utils.toSeconds(DEFAULT_BACK_BUFFER_DURATION_MS)));
+        long cacheMaxSize = (long)(options.getDouble("maxCacheSize", 0) * 1024);
         int multiplier = DEFAULT_BUFFER_FOR_PLAYBACK_AFTER_REBUFFER_MS / DEFAULT_BUFFER_FOR_PLAYBACK_MS;
 
         LoadControl control = new DefaultLoadControl.Builder()
                 .setBufferDurationsMs(minBuffer, maxBuffer, playBuffer, playBuffer * multiplier)
+                .setBackBuffer(backBuffer, false)
                 .createDefaultLoadControl();
 
-        SimpleExoPlayer player = ExoPlayerFactory.newSimpleInstance(new DefaultRenderersFactory(service), new DefaultTrackSelector(), control);
+        SimpleExoPlayer player = ExoPlayerFactory.newSimpleInstance(service, new DefaultRenderersFactory(service), new DefaultTrackSelector(), control);
 
         player.setAudioAttributes(new com.google.android.exoplayer2.audio.AudioAttributes.Builder()
                 .setContentType(C.CONTENT_TYPE_MUSIC).setUsage(C.USAGE_MEDIA).build());
 
-        return new ExoPlayback(service, this, player, cacheMaxSize);
+        return new LocalPlayback(service, this, player, cacheMaxSize);
     }
 
+    @SuppressLint("WakelockTimeout")
     public void onPlay() {
         Log.d(Utils.LOG, "onPlay");
-        if (playback == null) return;
+        if(playback == null) return;
 
         Track track = playback.getCurrentTrack();
-        if (track == null) return;
+        if(track == null) return;
 
-        if (!playback.isRemote()) {
+        if(!playback.isRemote()) {
             requestFocus();
 
-            if (!wakeLock.isHeld()) {
-                long timeout = playback.getDuration() - playback.getPosition();
-                wakeLock.acquire(timeout <= 0 ? (5 * 60 * 1000) : (timeout + 5000));
+            if(!receivingNoisyEvents) {
+                service.registerReceiver(noisyReceiver, new IntentFilter(AudioManager.ACTION_AUDIO_BECOMING_NOISY));
+                receivingNoisyEvents = true;
             }
 
-            if (!Utils.isLocal(track.uri)) {
-                if (!wifiLock.isHeld()) wifiLock.acquire();
+            if(!wakeLock.isHeld()) wakeLock.acquire();
+
+            if(!Utils.isLocal(track.uri)) {
+                if(!wifiLock.isHeld()) wifiLock.acquire();
             }
         }
 
-//        metadata.setForeground(true, true);
+        metadata.setActive(true);
     }
 
     public void onPause() {
         Log.d(Utils.LOG, "onPause");
 
+        // Unregisters the noisy receiver
+        if(receivingNoisyEvents) {
+            service.unregisterReceiver(noisyReceiver);
+            receivingNoisyEvents = false;
+        }
+
         // Release the wake and the wifi locks
-        if (wakeLock.isHeld()) wakeLock.release();
-        if (wifiLock.isHeld()) wifiLock.release();
+        if(wakeLock.isHeld()) wakeLock.release();
+        if(wifiLock.isHeld()) wifiLock.release();
 
-        abandonFocus();
-
-//        metadata.setForeground(false, true);
+        metadata.setActive(true);
     }
 
     public void onStop() {
         Log.d(Utils.LOG, "onStop");
 
         // Release the wake and the wifi locks
-        if (wakeLock.isHeld()) wakeLock.release();
-        if (wifiLock.isHeld()) wifiLock.release();
+        if(wakeLock.isHeld()) wakeLock.release();
+        if(wifiLock.isHeld()) wifiLock.release();
 
         abandonFocus();
 
-//        metadata.setForeground(false, false);
+        metadata.setActive(false);
     }
 
     public void onStateChange(int state) {
@@ -149,13 +184,13 @@ public class MusicManager implements OnAudioFocusChangeListener {
         Bundle bundle = new Bundle();
         bundle.putInt("state", state);
         service.emit(MusicEvents.PLAYBACK_STATE, bundle);
-//        metadata.updatePlayback(playback);
+        metadata.updatePlayback(playback);
     }
 
     public void onTrackUpdate(Track previous, long prevPos, Track next) {
         Log.d(Utils.LOG, "onTrackUpdate");
 
-//        if (next != null) metadata.updateMetadata(next);
+        if(next != null) metadata.updateMetadata(next);
 
         Bundle bundle = new Bundle();
         bundle.putString("track", previous != null ? previous.id : null);
@@ -165,7 +200,7 @@ public class MusicManager implements OnAudioFocusChangeListener {
     }
 
     public void onReset() {
-//        metadata.removeNotifications();
+        metadata.removeNotifications();
     }
 
     public void onEnd(Track previous, long prevPos) {
@@ -191,37 +226,41 @@ public class MusicManager implements OnAudioFocusChangeListener {
     public void onAudioFocusChange(int focus) {
         Log.d(Utils.LOG, "onDuck");
 
+        boolean permanent = false;
         boolean paused = false;
         boolean ducking = false;
 
-        switch (focus) {
+        switch(focus) {
             case AudioManager.AUDIOFOCUS_LOSS:
+                permanent = true;
+                abandonFocus();
             case AudioManager.AUDIOFOCUS_LOSS_TRANSIENT:
                 paused = true;
                 break;
             case AudioManager.AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK:
                 ducking = true;
                 break;
-            case AudioManager.AUDIOFOCUS_GAIN:
+            default:
                 break;
         }
 
         Bundle bundle = new Bundle();
+        bundle.putBoolean("permanent", permanent);
         bundle.putBoolean("paused", paused);
         bundle.putBoolean("ducking", ducking);
         service.emit(MusicEvents.BUTTON_DUCK, bundle);
     }
 
     private void requestFocus() {
-        if (hasAudioFocus) return;
+        if(hasAudioFocus) return;
         Log.d(Utils.LOG, "Requesting audio focus...");
 
-        AudioManager manager = (AudioManager) service.getSystemService(Context.AUDIO_SERVICE);
+        AudioManager manager = (AudioManager)service.getSystemService(Context.AUDIO_SERVICE);
         int r;
 
-        if (manager == null) {
+        if(manager == null) {
             r = AudioManager.AUDIOFOCUS_REQUEST_FAILED;
-        } else if (Build.VERSION.SDK_INT >= 26) {
+        } else if(Build.VERSION.SDK_INT >= 26) {
             focus = new AudioFocusRequest.Builder(AudioManager.AUDIOFOCUS_GAIN)
                     .setOnAudioFocusChangeListener(this)
                     .setAudioAttributes(new AudioAttributes.Builder()
@@ -240,15 +279,15 @@ public class MusicManager implements OnAudioFocusChangeListener {
     }
 
     private void abandonFocus() {
-        if (!hasAudioFocus) return;
+        if(!hasAudioFocus) return;
         Log.d(Utils.LOG, "Abandoning audio focus...");
 
-        AudioManager manager = (AudioManager) service.getSystemService(Context.AUDIO_SERVICE);
+        AudioManager manager = (AudioManager)service.getSystemService(Context.AUDIO_SERVICE);
         int r;
 
-        if (manager == null) {
+        if(manager == null) {
             r = AudioManager.AUDIOFOCUS_REQUEST_FAILED;
-        } else if (Build.VERSION.SDK_INT >= 26) {
+        } else if(Build.VERSION.SDK_INT >= 26) {
             r = manager.abandonAudioFocusRequest(focus);
         } else {
             //noinspection deprecation
@@ -264,14 +303,20 @@ public class MusicManager implements OnAudioFocusChangeListener {
         // Disable audio focus
         abandonFocus();
 
+        // Stop receiving audio becoming noisy events
+        if(receivingNoisyEvents) {
+            service.unregisterReceiver(noisyReceiver);
+            receivingNoisyEvents = false;
+        }
+
         // Release the playback resources
-        if (playback != null) playback.destroy();
+        if(playback != null) playback.destroy();
 
         // Release the metadata resources
-//        metadata.destroy();
+        metadata.destroy();
 
         // Release the locks
-        if (wifiLock.isHeld()) wifiLock.release();
-        if (wakeLock.isHeld()) wakeLock.release();
+        if(wifiLock.isHeld()) wifiLock.release();
+        if(wakeLock.isHeld()) wakeLock.release();
     }
 }
