@@ -1,15 +1,22 @@
 package com.guichaguri.trackplayer.service;
 
+import android.app.Activity;
 import android.app.Notification;
+import android.app.NotificationChannel;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
 import android.os.IBinder;
+import android.support.v4.app.NotificationCompat;
 import android.support.v4.content.LocalBroadcastManager;
 import android.support.v4.media.session.MediaButtonReceiver;
 
+import android.support.v4.media.session.MediaSessionCompat;
 import com.facebook.react.HeadlessJsTaskService;
+import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.bridge.Arguments;
+import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.jstasks.HeadlessJsTaskConfig;
 import javax.annotation.Nullable;
 
@@ -18,15 +25,8 @@ import javax.annotation.Nullable;
  */
 public class MusicService extends HeadlessJsTaskService {
 
-    private MusicManager manager;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            startForeground(1, new Notification());
-        }
-    }
+    MusicManager manager;
+    Handler handler;
 
     @Nullable
     @Override
@@ -48,11 +48,51 @@ public class MusicService extends HeadlessJsTaskService {
         LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
     }
 
+    public void destroy() {
+        if(handler != null) {
+            handler.removeMessages(0);
+            handler = null;
+        }
+
+        if(manager != null) {
+            manager.destroy();
+            manager = null;
+        }
+    }
+
+    private void onStartForeground() {
+        boolean serviceForeground = false;
+
+        if(manager != null) {
+            // The session is only active when the service is on foreground
+            serviceForeground = manager.getMetadata().getSession().isActive();
+        }
+
+        if(!serviceForeground) {
+            ReactInstanceManager reactInstanceManager = getReactNativeHost().getReactInstanceManager();
+            ReactContext reactContext = reactInstanceManager.getCurrentReactContext();
+
+            // Checks whether there is a React activity
+            if(reactContext == null || !reactContext.hasCurrentActivity()) {
+                String channel = null;
+
+                if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                    channel = NotificationChannel.DEFAULT_CHANNEL_ID;
+                }
+
+                // Sets the service to foreground with an empty notification
+                startForeground(1, new NotificationCompat.Builder(this, channel).build());
+                // Stops the service right after
+                stopSelf();
+            }
+        }
+    }
+
     @Nullable
     @Override
     public IBinder onBind(Intent intent) {
         if(Utils.CONNECT_INTENT.equals(intent.getAction())) {
-            return new MusicBinder(manager);
+            return new MusicBinder(this, manager);
         }
 
         return super.onBind(intent);
@@ -60,12 +100,20 @@ public class MusicService extends HeadlessJsTaskService {
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
-        if (intent != null && "android.intent.action.MEDIA_BUTTON".equals(intent.getAction())) {
-        //    MediaButtonReceiver.handleIntent(manager.getMetadata().getSession(), intent);
-              return START_NOT_STICKY;
+        if(intent != null && Intent.ACTION_MEDIA_BUTTON.equals(intent.getAction())) {
+            // Check if the app is on background, then starts a foreground service and then ends it right after
+            onStartForeground();
+
+            if(manager != null) {
+                MediaButtonReceiver.handleIntent(manager.getMetadata().getSession(), intent);
+            }
+
+            return START_NOT_STICKY;
         }
 
         manager = new MusicManager(this);
+        handler = new Handler();
+
         super.onStartCommand(intent, flags, startId);
         return START_STICKY;
     }
@@ -74,7 +122,15 @@ public class MusicService extends HeadlessJsTaskService {
     public void onDestroy() {
         super.onDestroy();
 
-        manager.destroy();
-        manager = null;
+        destroy();
+    }
+
+    @Override
+    public void onTaskRemoved(Intent rootIntent) {
+        super.onTaskRemoved(rootIntent);
+
+        if (manager == null || manager.shouldStopWithApp()) {
+            stopSelf();
+        }
     }
 }
