@@ -1,6 +1,8 @@
 package com.guichaguri.trackplayer.service.metadata;
 
+import android.annotation.SuppressLint;
 import android.os.Bundle;
+import android.provider.MediaStore;
 import android.support.v4.media.RatingCompat;
 import android.support.v4.media.session.MediaSessionCompat;
 import com.guichaguri.trackplayer.module.MusicEvents;
@@ -8,6 +10,7 @@ import com.guichaguri.trackplayer.service.MusicManager;
 import com.guichaguri.trackplayer.service.MusicService;
 import com.guichaguri.trackplayer.service.Utils;
 import com.guichaguri.trackplayer.service.models.Track;
+import java.util.List;
 
 /**
  * @author Guichaguri
@@ -37,9 +40,57 @@ public class ButtonEvents extends MediaSessionCompat.Callback {
         service.emit(MusicEvents.BUTTON_STOP, null);
     }
 
+
+    @Override
+    public void onPlayFromMediaId(String mediaId, Bundle extras) {
+        Bundle bundle = new Bundle();
+        bundle.putString("id", mediaId);
+        service.emit(MusicEvents.BUTTON_PLAY_FROM_ID, bundle);
+    }
+
+    @SuppressLint("InlinedApi")
+    @Override
+    public void onPlayFromSearch(String query, Bundle extras) {
+        Bundle bundle = new Bundle();
+        bundle.putString("query", query);
+
+        if(extras.containsKey(MediaStore.EXTRA_MEDIA_FOCUS)) {
+            String focus = extras.getString(MediaStore.EXTRA_MEDIA_FOCUS);
+
+            if(MediaStore.Audio.Artists.ENTRY_CONTENT_TYPE.equals(focus)) {
+                focus = "artist";
+            } else if(MediaStore.Audio.Albums.ENTRY_CONTENT_TYPE.equals(focus)) {
+                focus = "album";
+            } else if(MediaStore.Audio.Playlists.ENTRY_CONTENT_TYPE.equals(focus)) {
+                focus = "playlist";
+            } else if(MediaStore.Audio.Genres.ENTRY_CONTENT_TYPE.equals(focus)) {
+                focus = "genre";
+            } else if(MediaStore.Audio.Media.ENTRY_CONTENT_TYPE.equals(focus)) {
+                focus = "title";
+            }
+
+            bundle.putString("focus", focus);
+        }
+
+        if(extras.containsKey(MediaStore.EXTRA_MEDIA_TITLE))
+            bundle.putString("title", extras.getString(MediaStore.EXTRA_MEDIA_TITLE));
+        if(extras.containsKey(MediaStore.EXTRA_MEDIA_ARTIST))
+            bundle.putString("artist", extras.getString(MediaStore.EXTRA_MEDIA_ARTIST));
+        if(extras.containsKey(MediaStore.EXTRA_MEDIA_ALBUM))
+            bundle.putString("album", extras.getString(MediaStore.EXTRA_MEDIA_ALBUM));
+        if(extras.containsKey(MediaStore.EXTRA_MEDIA_GENRE))
+            bundle.putString("genre", extras.getString(MediaStore.EXTRA_MEDIA_GENRE));
+        if(extras.containsKey(MediaStore.EXTRA_MEDIA_PLAYLIST))
+            bundle.putString("playlist", extras.getString(MediaStore.EXTRA_MEDIA_PLAYLIST));
+
+        service.emit(MusicEvents.BUTTON_PLAY_FROM_SEARCH, bundle);
+    }
+
     @Override
     public void onSkipToQueueItem(long id) {
-        for(Track track : manager.getPlayback().getQueue()) {
+        List<Track> tracks = manager.getPlayback().getQueue();
+
+        for(Track track : tracks) {
             if(track.queueId != id) continue;
 
             Bundle bundle = new Bundle();
