@@ -1,25 +1,24 @@
 package com.guichaguri.trackplayer.service.metadata;
 
-import android.app.Notification;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Bitmap;
+import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.NotificationCompat;
 import android.support.v4.app.NotificationCompat.Action;
-import android.support.v4.app.NotificationManagerCompat;
 import android.support.v4.media.MediaMetadataCompat;
 import android.support.v4.media.RatingCompat;
 import android.support.v4.media.app.NotificationCompat.MediaStyle;
 import android.support.v4.media.session.MediaButtonReceiver;
 import android.support.v4.media.session.MediaSessionCompat;
 import android.support.v4.media.session.PlaybackStateCompat;
-
 import com.bumptech.glide.Glide;
+import com.bumptech.glide.RequestManager;
 import com.bumptech.glide.request.target.SimpleTarget;
 import com.bumptech.glide.request.transition.Transition;
 import com.facebook.react.views.imagehelper.ResourceDrawableIdHelper;
@@ -29,7 +28,6 @@ import com.guichaguri.trackplayer.service.MusicService;
 import com.guichaguri.trackplayer.service.Utils;
 import com.guichaguri.trackplayer.service.models.Track;
 import com.guichaguri.trackplayer.service.player.ExoPlayback;
-
 import java.util.ArrayList;
 import java.util.List;
 
@@ -42,12 +40,12 @@ public class MetadataManager {
     private final MusicManager manager;
     private final MediaSessionCompat session;
 
-    private boolean foreground = false;
     private int ratingType = RatingCompat.RATING_NONE;
+    private int jumpInterval = 15;
     private long actions = 0;
     private long compactActions = 0;
+    private SimpleTarget<Bitmap> artworkTarget;
     private NotificationCompat.Builder builder;
-    private String notificationChannel = "trackplayer";
 
     private Action previousAction, rewindAction, playAction, pauseAction, stopAction, forwardAction, nextAction;
 
@@ -56,7 +54,7 @@ public class MetadataManager {
         this.manager = manager;
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            NotificationChannel channel = new NotificationChannel(notificationChannel, "notification service", NotificationManager.IMPORTANCE_DEFAULT);
+            NotificationChannel channel = new NotificationChannel(Utils.NOTIFICATION_CHANNEL, "Playback", NotificationManager.IMPORTANCE_DEFAULT);
             channel.setShowBadge(false);
             channel.setSound(null, null);
 
@@ -64,7 +62,7 @@ public class MetadataManager {
             not.createNotificationChannel(channel);
         }
 
-        this.builder = new NotificationCompat.Builder(service, notificationChannel);
+        this.builder = new NotificationCompat.Builder(service, Utils.NOTIFICATION_CHANNEL);
         this.session = new MediaSessionCompat(service, "TrackPlayer", null, null);
 
         session.setFlags(MediaSessionCompat.FLAG_HANDLES_MEDIA_BUTTONS |
@@ -74,7 +72,15 @@ public class MetadataManager {
         Context context = service.getApplicationContext();
         String packageName = context.getPackageName();
         Intent openApp = context.getPackageManager().getLaunchIntentForPackage(packageName);
-        builder.setContentIntent(PendingIntent.getActivity(context, 0, openApp, 0));
+
+        // Prevent the app from launching a new instance
+        openApp.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_SINGLE_TOP);
+
+        // Add the Uri data so apps can identify that it was a notification click
+        openApp.setAction(Intent.ACTION_VIEW);
+        openApp.setData(Uri.parse("trackplayer://notification.click"));
+
+        builder.setContentIntent(PendingIntent.getActivity(context, 0, openApp, PendingIntent.FLAG_CANCEL_CURRENT));
 
         builder.setSmallIcon(R.drawable.play);
         builder.setCategory(NotificationCompat.CATEGORY_TRANSPORT);
@@ -92,7 +98,6 @@ public class MetadataManager {
 
     /**
      * Updates the metadata options
-     *
      * @param options The options
      */
     public void updateOptions(Bundle options) {
@@ -103,12 +108,12 @@ public class MetadataManager {
         actions = 0;
         compactActions = 0;
 
-        if (capabilities != null) {
+        if(capabilities != null) {
             // Create the actions mask
-            for (int cap : capabilities) actions |= cap;
+            for(int cap : capabilities) actions |= cap;
 
             // If there is no notification capabilities defined, we'll show all capabilities available
-            if (notification == null) notification = capabilities;
+            if(notification == null) notification = capabilities;
 
             // Initialize all actions based on the options
 
@@ -128,26 +133,33 @@ public class MetadataManager {
                     getIcon(options, "nextIcon", R.drawable.next));
 
             // Update the action mask for the compact view
-            if (compact != null) {
-                for (int cap : compact) compactActions |= cap;
+            if(compact != null) {
+                for(int cap : compact) compactActions |= cap;
             }
         }
 
-//        // Update the color
+        // Update the color
         builder.setColor(options.getInt("color", NotificationCompat.COLOR_DEFAULT));
 
         // Update the icon
         builder.setSmallIcon(getIcon(options, "icon", R.drawable.play));
 
+        // Update the jump interval
+        jumpInterval = options.getInt("jumpInterval", 15);
+
         // Update the rating type
         ratingType = options.getInt("ratingType", RatingCompat.RATING_NONE);
         session.setRatingType(ratingType);
 
-//        updateNotification();
+        updateNotification();
     }
 
     public int getRatingType() {
         return ratingType;
+    }
+
+    public int getJumpInterval() {
+        return jumpInterval;
     }
 
     public void removeNotifications() {
@@ -159,12 +171,11 @@ public class MetadataManager {
 
     /**
      * Updates the artwork
-     *
      * @param bitmap The new artwork
      */
     protected void updateArtwork(Bitmap bitmap) {
         Track track = manager.getPlayback().getCurrentTrack();
-        if (track == null) return;
+        if(track == null) return;
 
         MediaMetadataCompat.Builder metadata = track.toMediaMetadata();
 
@@ -177,15 +188,16 @@ public class MetadataManager {
 
     /**
      * Updates the current track
-     *
      * @param track The new track
      */
-    public void updateMetadata(final Track track) {
-        final MediaMetadataCompat.Builder metadata = track.toMediaMetadata();
+    public void updateMetadata(Track track) {
+        MediaMetadataCompat.Builder metadata = track.toMediaMetadata();
 
-        if (track.artwork != null) {
-            Glide.with(service.getApplicationContext())
-                    .asBitmap()
+        RequestManager rm = Glide.with(service.getApplicationContext());
+        if(artworkTarget != null) rm.clear(artworkTarget);
+
+        if(track.artwork != null) {
+            artworkTarget = rm.asBitmap()
                     .load(track.artwork)
                     .into(new SimpleTarget<Bitmap>() {
                         @Override
@@ -193,12 +205,9 @@ public class MetadataManager {
                             metadata.putBitmap(MediaMetadataCompat.METADATA_KEY_ART, resource);
                             builder.setLargeIcon(resource);
 
-                            builder.setContentTitle(track.title);
-                            builder.setContentText(track.artist);
-                            builder.setSubText(track.album);
-
                             session.setMetadata(metadata.build());
                             updateNotification();
+                            artworkTarget = null;
                         }
                     });
         }
@@ -213,13 +222,11 @@ public class MetadataManager {
 
     /**
      * Updates the playback state
-     *
      * @param playback The player
      */
     public void updatePlayback(ExoPlayback playback) {
         int state = playback.getState();
         boolean playing = Utils.isPlaying(state);
-        MediaStyle style = new MediaStyle();
         List<Integer> compact = new ArrayList<>();
         builder.mActions.clear();
 
@@ -228,34 +235,45 @@ public class MetadataManager {
         addAction(previousAction, PlaybackStateCompat.ACTION_SKIP_TO_PREVIOUS, compact);
         addAction(rewindAction, PlaybackStateCompat.ACTION_REWIND, compact);
 
-        if (playing) {
+        if(playing) {
             addAction(pauseAction, PlaybackStateCompat.ACTION_PAUSE, compact);
-            style.setShowCancelButton(false);
         } else {
             addAction(playAction, PlaybackStateCompat.ACTION_PLAY, compact);
-
-            // Shows the cancel button on pre-lollipop versions due to a bug
-            style.setShowCancelButton(true);
-            style.setCancelButtonIntent(MediaButtonReceiver.buildMediaButtonPendingIntent(service, PlaybackStateCompat.ACTION_STOP));
         }
 
         addAction(stopAction, PlaybackStateCompat.ACTION_STOP, compact);
         addAction(forwardAction, PlaybackStateCompat.ACTION_FAST_FORWARD, compact);
         addAction(nextAction, PlaybackStateCompat.ACTION_SKIP_TO_NEXT, compact);
 
-        // Links the media session
-        style.setMediaSession(session.getSessionToken());
+        // Prevent the media style from being used in older Huawei devices that don't support custom styles
+        if(!Build.MANUFACTURER.toLowerCase().contains("huawei") || Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
 
-        // Updates the compact media buttons for the notification
-        if (!compact.isEmpty()) {
-            int[] compactIndexes = new int[compact.size()];
+            MediaStyle style = new MediaStyle();
 
-            for (int i = 0; i < compact.size(); i++) compactIndexes[i] = compact.get(i);
+            if(playing) {
+                style.setShowCancelButton(false);
+            } else {
+                // Shows the cancel button on pre-lollipop versions due to a bug
+                style.setShowCancelButton(true);
+                style.setCancelButtonIntent(MediaButtonReceiver.buildMediaButtonPendingIntent(service,
+                        PlaybackStateCompat.ACTION_STOP));
+            }
 
-            style.setShowActionsInCompactView(compactIndexes);
+            // Links the media session
+            style.setMediaSession(session.getSessionToken());
+
+            // Updates the compact media buttons for the notification
+            if (!compact.isEmpty()) {
+                int[] compactIndexes = new int[compact.size()];
+
+                for (int i = 0; i < compact.size(); i++) compactIndexes[i] = compact.get(i);
+
+                style.setShowActionsInCompactView(compactIndexes);
+            }
+
+            builder.setStyle(style);
+
         }
-
-        builder.setStyle(style);
 
         // Updates the media session state
         PlaybackStateCompat.Builder pb = new PlaybackStateCompat.Builder();
@@ -267,70 +285,50 @@ public class MetadataManager {
         updateNotification();
     }
 
-    public void setForeground(boolean foreground, boolean active) {
-        this.foreground = foreground;
+    public void setActive(boolean active) {
         this.session.setActive(active);
 
-        if (foreground) {
-            updateNotification();
-        } else {
-            service.stopForeground(false);
-        }
+        updateNotification();
     }
 
     public void destroy() {
-        if (foreground) {
-            service.stopForeground(true);
-        } else {
-            NotificationManagerCompat.from(service).cancel(1);
-        }
+        service.stopForeground(true);
 
         session.setActive(false);
         session.release();
     }
 
     private void updateNotification() {
-        if (manager.getPlayback() == null) {
-            return;
-        }
-        int state = manager.getPlayback().getState();
-//        if (Utils.isStopped(state)) {
-//            removeNotifications();
-//            return;
-//        }
-
-        Notification n = builder.build();
-
-        if (foreground) {
-            service.startForeground(1, n);
+        if(session.isActive()) {
+            service.startForeground(1, builder.build());
         } else {
-            NotificationManagerCompat.from(service).notify(1, n);
+            service.stopForeground(true);
         }
     }
 
     private int getIcon(Bundle options, String propertyName, int defaultIcon) {
-        if (!options.containsKey(propertyName)) return defaultIcon;
+        if(!options.containsKey(propertyName)) return defaultIcon;
 
         Bundle bundle = options.getBundle(propertyName);
-        if (bundle == null) return defaultIcon;
+        if(bundle == null) return defaultIcon;
 
         ResourceDrawableIdHelper helper = ResourceDrawableIdHelper.getInstance();
         int icon = helper.getResourceDrawableId(service, bundle.getString("uri"));
-        if (icon == 0) return defaultIcon;
+        if(icon == 0) return defaultIcon;
 
         return icon;
     }
 
     private Action createAction(List<Integer> caps, long action, String title, int icon) {
-        if (!caps.contains((int) action)) return null;
+        if(!caps.contains((int)action)) return null;
 
         return new Action(icon, title, MediaButtonReceiver.buildMediaButtonPendingIntent(service, action));
     }
 
     private void addAction(Action action, long id, List<Integer> compact) {
-        if (action == null) return;
+        if(action == null) return;
 
-        if ((compactActions & id) != 0) compact.add(builder.mActions.size());
+        if((compactActions & id) != 0) compact.add(builder.mActions.size());
         builder.mActions.add(action);
     }
 
