@@ -9,22 +9,13 @@ import com.google.android.exoplayer2.ExoPlaybackException;
 import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.Player.EventListener;
-import com.google.android.exoplayer2.SimpleExoPlayer;
 import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.Timeline.Window;
-import com.google.android.exoplayer2.source.ConcatenatingMediaSource;
-import com.google.android.exoplayer2.source.MediaSource;
 import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
-import com.google.android.exoplayer2.upstream.DataSource;
-import com.google.android.exoplayer2.upstream.cache.CacheDataSource;
-import com.google.android.exoplayer2.upstream.cache.CacheDataSourceFactory;
-import com.google.android.exoplayer2.upstream.cache.LeastRecentlyUsedCacheEvictor;
-import com.google.android.exoplayer2.upstream.cache.SimpleCache;
 import com.guichaguri.trackplayer.service.MusicManager;
 import com.guichaguri.trackplayer.service.Utils;
 import com.guichaguri.trackplayer.service.models.Track;
-import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -33,110 +24,40 @@ import java.util.List;
 /**
  * @author Guichaguri
  */
-public class ExoPlayback implements EventListener {
+public abstract class ExoPlayback<T extends Player> implements EventListener {
 
-    private final Context context;
-    private final MusicManager manager;
-    private final SimpleExoPlayer player;
-    private final SimpleCache cache;
-    private final long cacheMaxSize;
+    protected final Context context;
+    protected final MusicManager manager;
+    protected final T player;
 
-    private ConcatenatingMediaSource source;
-    private List<Track> queue = Collections.synchronizedList(new ArrayList<Track>());
+    protected List<Track> queue = Collections.synchronizedList(new ArrayList<>());
 
     // https://github.com/google/ExoPlayer/issues/2728
-    private int lastKnownWindow = C.INDEX_UNSET;
-    private long lastKnownPosition = C.POSITION_UNSET;
-    private int previousState = PlaybackStateCompat.STATE_NONE;
+    protected int lastKnownWindow = C.INDEX_UNSET;
+    protected long lastKnownPosition = C.POSITION_UNSET;
+    protected int previousState = PlaybackStateCompat.STATE_NONE;
 
-    public ExoPlayback(Context context, MusicManager manager, SimpleExoPlayer player, long maxCacheSize) {
+    public ExoPlayback(Context context, MusicManager manager, T player) {
         this.context = context;
         this.manager = manager;
         this.player = player;
-        this.cacheMaxSize = maxCacheSize;
+    }
 
-        if(cacheMaxSize > 0) {
-            File cacheDir = new File(context.getCacheDir(), "TrackPlayer");
-            cache = new SimpleCache(cacheDir, new LeastRecentlyUsedCacheEvictor(cacheMaxSize));
-        } else {
-            cache = null;
-        }
-
+    public void initialize() {
         player.addListener(this);
-        resetQueue();
-    }
-
-    private void resetQueue() {
-        queue.clear();
-
-        source = new ConcatenatingMediaSource();
-        player.prepare(source);
-
-        lastKnownWindow = C.INDEX_UNSET;
-        lastKnownPosition = C.POSITION_UNSET;
-
-        manager.onReset();
-    }
-
-    public DataSource.Factory enableCaching(DataSource.Factory ds) {
-        if(cache == null || cacheMaxSize <= 0) return ds;
-
-        return new CacheDataSourceFactory(cache, ds, CacheDataSource.FLAG_IGNORE_CACHE_ON_ERROR, cacheMaxSize);
     }
 
     public List<Track> getQueue() {
         return queue;
     }
 
-    public void add(Track track, int index, Promise promise) {
-        queue.add(index, track);
-        source.addMediaSource(index, track.toMediaSource(context, cacheMaxSize), Utils.toRunnable(promise));
+    public abstract void add(Track track, int index, Promise promise);
 
-        if (queue.size() == 1) {
-            player.prepare(source);
-        }
-    }
+    public abstract void add(Collection<Track> tracks, int index, Promise promise);
 
-    public void add(Collection<Track> tracks, int index, Promise promise) {
-        List<MediaSource> trackList = new ArrayList<>();
+    public abstract void remove(List<Integer> indexes, Promise promise);
 
-        for(Track track : tracks) {
-            trackList.add(track.toMediaSource(context, cacheMaxSize));
-        }
-
-        queue.addAll(index, tracks);
-        source.addMediaSources(index, trackList, Utils.toRunnable(promise));
-
-        if (queue.size() == tracks.size()) {
-            player.prepare(source);
-        }
-    }
-
-    public void remove(List<Integer> indexes, Promise promise) {
-        Collections.sort(indexes);
-
-        for(int i = indexes.size() - 1; i >= 0; i--) {
-            int index = indexes.get(i);
-
-            queue.remove(index);
-
-            if(i == 0) {
-                source.removeMediaSource(index, Utils.toRunnable(promise));
-            } else {
-                source.removeMediaSource(index, null);
-            }
-        }
-    }
-
-    public void removeUpcomingTracks() {
-        int currentIndex = player.getCurrentWindowIndex();
-        if (currentIndex == C.INDEX_UNSET) return;
-
-        for (int i = queue.size() - 1; i > currentIndex; i--) {
-            queue.remove(i);
-            source.removeMediaSource(i, null);
-        }
-    }
+    public abstract void removeUpcomingTracks();
 
     public Track getCurrentTrack() {
         int index = player.getCurrentWindowIndex();
@@ -144,6 +65,11 @@ public class ExoPlayback implements EventListener {
     }
 
     public void skip(String id, Promise promise) {
+        if(id == null || id.isEmpty()) {
+            promise.reject("invalid_id", "The ID can't be null or empty");
+            return;
+        }
+
         for(int i = 0; i < queue.size(); i++) {
             if(id.equals(queue.get(i).id)) {
                 lastKnownWindow = player.getCurrentWindowIndex();
@@ -201,7 +127,8 @@ public class ExoPlayback implements EventListener {
         lastKnownPosition = player.getCurrentPosition();
 
         player.stop(false);
-        player.seekTo(0);
+        player.setPlayWhenReady(false);
+        player.seekTo(0,0);
     }
 
     public void reset() {
@@ -209,7 +136,7 @@ public class ExoPlayback implements EventListener {
         lastKnownPosition = player.getCurrentPosition();
 
         player.stop(true);
-        resetQueue();
+        player.setPlayWhenReady(false);
     }
 
     public boolean isRemote() {
@@ -235,13 +162,9 @@ public class ExoPlayback implements EventListener {
         player.seekTo(time);
     }
 
-    public float getVolume() {
-        return player.getVolume();
-    }
+    public abstract float getVolume();
 
-    public void setVolume(float volume) {
-        player.setVolume(volume);
-    }
+    public abstract void setVolume(float volume);
 
     public float getRate() {
         return player.getPlaybackParameters().speed;
@@ -254,7 +177,7 @@ public class ExoPlayback implements EventListener {
     public int getState() {
         switch(player.getPlaybackState()) {
             case Player.STATE_BUFFERING:
-                return player.getPlayWhenReady() ? PlaybackStateCompat.STATE_BUFFERING : PlaybackStateCompat.STATE_NONE;
+                return PlaybackStateCompat.STATE_BUFFERING;
             case Player.STATE_ENDED:
                 return PlaybackStateCompat.STATE_STOPPED;
             case Player.STATE_IDLE:
@@ -273,7 +196,7 @@ public class ExoPlayback implements EventListener {
     public void onTimelineChanged(Timeline timeline, Object manifest, int reason) {
         Log.d(Utils.LOG, "onTimelineChanged: " + reason);
 
-        if ((reason == Player.TIMELINE_CHANGE_REASON_PREPARED || reason == Player.TIMELINE_CHANGE_REASON_DYNAMIC) && !timeline.isEmpty()) {
+        if((reason == Player.TIMELINE_CHANGE_REASON_PREPARED || reason == Player.TIMELINE_CHANGE_REASON_DYNAMIC) && !timeline.isEmpty()) {
             onPositionDiscontinuity(Player.DISCONTINUITY_REASON_INTERNAL);
         }
     }
@@ -345,7 +268,17 @@ public class ExoPlayback implements EventListener {
 
     @Override
     public void onPlayerError(ExoPlaybackException error) {
-        manager.onError("exoplayer", error.getCause().getMessage());
+        String code;
+
+        if(error.type == ExoPlaybackException.TYPE_SOURCE) {
+            code = "playback-source";
+        } else if(error.type == ExoPlaybackException.TYPE_RENDERER) {
+            code = "playback-renderer";
+        } else {
+            code = "playback"; // Other unexpected errors related to the playback
+        }
+
+        manager.onError(code, error.getCause().getMessage());
     }
 
     @Override
