@@ -9,5 +9,26 @@
 import Foundation
 
 enum Capability: String {
-    case play, pause, stop, next, previous, jumpForward, jumpBackward, seek, unsupported
+    case play, pause, stop, next, previous, jumpForward, jumpBackward, seek
+    
+    func mapToPlayerCommand(jumpInterval: NSNumber?) -> RemoteCommand {
+        switch self {
+        case .stop:
+            return .stop
+        case .play:
+            return .play
+        case .pause:
+            return .pause
+        case .next:
+            return .next
+        case .previous:
+            return .previous
+        case .seek:
+            return .changePlaybackPosition
+        case .jumpForward:
+            return .skipForward(preferredIntervals: [jumpInterval ?? 15])
+        case .jumpBackward:
+            return .skipBackward(preferredIntervals: [jumpInterval ?? 15])
+        }
+    }
 }
