@@ -18,14 +18,14 @@ struct MediaURL {
         originalObject = object
         
         if let localObject = object as? [String: Any] {
-            isLocal = true
             let uri = localObject["uri"] as! String
+            isLocal = uri.contains("http") ? false : true
             let encodedURI = uri.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed)!
-            value = URL(string: encodedURI)!
+            value = URL(string: encodedURI.replacingOccurrences(of: "file://", with: ""))!
         } else {
             let url = object as! String
             isLocal = url.lowercased().hasPrefix("file://")
-            value = URL(string: url)!
+            value = URL(string: url.replacingOccurrences(of: "file://", with: ""))!
         }
     }
 }
