@@ -2,6 +2,7 @@ import { Platform, AppRegistry, DeviceEventEmitter, NativeEventEmitter, NativeMo
 import resolveAssetSource from 'react-native/Libraries/Image/resolveAssetSource';
 
 const { TrackPlayerModule: TrackPlayer } = NativeModules;
+const emitter = Platform.OS !== 'android' ? new NativeEventEmitter(TrackPlayer) : DeviceEventEmitter;
 
 function resolveAsset(uri) {
     if(!uri) return undefined;
@@ -28,6 +29,8 @@ function updateOptions(data) {
     data.stopIcon = resolveAsset(data.stopIcon);
     data.previousIcon = resolveAsset(data.previousIcon);
     data.nextIcon = resolveAsset(data.nextIcon);
+    data.rewindIcon = resolveAsset(data.rewindIcon);
+    data.forwardIcon = resolveAsset(data.forwardIcon);
 
     return TrackPlayer.updateOptions(data);
 }
@@ -60,16 +63,30 @@ function remove(tracks) {
     return TrackPlayer.remove(tracks);
 }
 
+function registerPlaybackService(serviceFactory) {
+    if (Platform.OS === 'android') {
+        // Registers the headless task
+        AppRegistry.registerHeadlessTask('TrackPlayer', serviceFactory);
+    } else {
+        // Initializes and runs the service in the next tick
+        setImmediate(serviceFactory());
+    }
+}
+
+function addEventListener(event, listener) {
+    return emitter.addListener(event, listener);
+}
+
 function warpEventResponse(handler, event, payload) {
-    // transform into headlessTask format and return to handler
+    // transform into the old format and return to handler
     const additionalKeys = payload || {};
     handler({ type: event, ...additionalKeys });
 }
 
+/**
+ * @deprecated since version 1.0.1. Use addEventListener instead.
+ */
 function registerEventHandler(handler) {
-    const isAndroid = Platform.OS === 'android';
-    const emitter = !isAndroid ? new NativeEventEmitter(TrackPlayer) : DeviceEventEmitter;
-
     let events = [
         'playback-state',
         'playback-error',
@@ -85,18 +102,18 @@ function registerEventHandler(handler) {
         'remote-jump-backward',
         'remote-seek',
     ];
-
-    if (isAndroid) {
-        events.push('remote-skip', 'remote-duck', 'remote-set-rating');
+  
+    if (Platform.OS === 'android') {
+        events.push('remote-skip', 'remote-duck', 'remote-set-rating', 'remote-play-id', 'remote-play-search');
     }
 
-    for (let i = 0; i < events.length; i++) {
-        emitter.addListener(events[i], warpEventResponse.bind(null, handler, events[i]));
-    }
-
-    if (isAndroid) {
-        AppRegistry.registerHeadlessTask('TrackPlayer', () => handler);
-    }
+    registerPlaybackService(() => {
+        return async function() {
+            for (let i = 0; i < events.length; i++) {
+                addEventListener(events[i], warpEventResponse.bind(null, handler, events[i]));
+            }
+        };
+    });
 }
 
 // We'll declare each one of the constants and functions manually so IDEs can show a list of them
@@ -141,6 +158,9 @@ module.exports.setupPlayer = setupPlayer;
 module.exports.destroy = TrackPlayer.destroy;
 module.exports.updateOptions = updateOptions;
 module.exports.registerEventHandler = registerEventHandler;
+module.exports.registerBackendService = registerPlaybackService;
+module.exports.registerPlaybackService = registerPlaybackService;
+module.exports.addEventListener = addEventListener;
 
 // Player Queue Commands
 module.exports.add = add;
