@@ -1,14 +1,9 @@
 using Newtonsoft.Json.Linq;
 using ReactNative.Bridge;
-using ReactNative.Collections;
-using System;
 using System.Collections.Generic;
-using Windows.ApplicationModel.Core;
-using Windows.UI.Core;
-using Windows.UI.Popups;
-using Windows.Media.Playback;
 using TrackPlayer.Logic;
 using TrackPlayer.Players;
+using System.Threading.Tasks;
 
 namespace TrackPlayer
 {
@@ -16,24 +11,34 @@ namespace TrackPlayer
     {
         public override string Name => "TrackPlayerModule";
 
-        public override IReadOnlyDictionary<string, object> Constants => new Dictionary<string, object> {
-            {"STATE_NONE", (int)MediaPlaybackState.None},
-            {"STATE_PLAYING", (int)MediaPlaybackState.Playing},
-            {"STATE_PAUSED", (int)MediaPlaybackState.Paused},
-            {"STATE_STOPPED", -1}, // Unsupported
-            {"STATE_BUFFERING", (int)MediaPlaybackState.Buffering},
+        public override JObject ModuleConstants {
+            get {
+                var obj = new JObject();
 
-            {"CAPABILITY_PLAY", (int)Capability.Play},
-            {"CAPABILITY_PLAY_FROM_ID", (int)Capability.Unsupported},
-            {"CAPABILITY_PLAY_FROM_SEARCH", (int)Capability.Unsupported},
-            {"CAPABILITY_PAUSE", (int)Capability.Pause},
-            {"CAPABILITY_STOP", (int)Capability.Stop},
-            {"CAPABILITY_SEEK_TO", (int)Capability.Seek},
-            {"CAPABILITY_SKIP", (int)Capability.Unsupported},
-            {"CAPABILITY_SKIP_TO_NEXT", (int)Capability.Next},
-            {"CAPABILITY_SKIP_TO_PREVIOUS", (int)Capability.Previous},
-            {"CAPABILITY_SET_RATING", (int)Capability.Unsupported}
-        };
+                // States
+                obj["STATE_NONE"] = (int) PlaybackState.None;
+                obj["STATE_PLAYING"] = (int) PlaybackState.Playing;
+                obj["STATE_PAUSED"] = (int) PlaybackState.Paused;
+                obj["STATE_STOPPED"] = (int) PlaybackState.Stopped;
+                obj["STATE_BUFFERING"] = (int) PlaybackState.Buffering;
+
+                // Capabilities
+                obj["CAPABILITY_PLAY"] = (int) Capability.Play;
+                obj["CAPABILITY_PLAY_FROM_ID"] = (int) Capability.Unsupported;
+                obj["CAPABILITY_PLAY_FROM_SEARCH"] = (int) Capability.Unsupported;
+                obj["CAPABILITY_PAUSE"] = (int) Capability.Pause;
+                obj["CAPABILITY_STOP"] = (int) Capability.Stop;
+                obj["CAPABILITY_SEEK_TO"] = (int) Capability.Seek;
+                obj["CAPABILITY_SKIP"] = (int) Capability.Unsupported;
+                obj["CAPABILITY_SKIP_TO_NEXT"] = (int) Capability.Next;
+                obj["CAPABILITY_SKIP_TO_PREVIOUS"] = (int) Capability.Previous;
+                obj["CAPABILITY_SET_RATING"] = (int) Capability.Unsupported;
+                obj["CAPABILITY_JUMP_FORWARD"] = (int) Capability.JumpForward;
+                obj["CAPABILITY_JUMP_BACKWARD"] = (int) Capability.JumpBackward;
+
+                return obj;
+            }
+        }
 
         private MediaManager manager;
 
@@ -42,79 +47,105 @@ namespace TrackPlayer
 
         }
 
-        public override void OnReactInstanceDispose()
+        public override void Initialize()
         {
-            base.OnReactInstanceDispose();
+            manager = new MediaManager(Context);
+            base.Initialize();
+        }
 
+        public override async Task OnReactInstanceDisposeAsync()
+        {
             if (manager != null)
             {
                 manager.Dispose();
                 manager = null;
             }
+
+            await base.OnReactInstanceDisposeAsync();
         }
 
         [ReactMethod]
         public void setupPlayer(JObject options, IPromise promise)
         {
-            if (manager != null)
+            if (manager.GetPlayer() != null)
             {
                 promise.Resolve(null);
                 return;
             }
 
-            manager = new MediaManager(Context, options);
+            manager.SwitchPlayback(manager.CreateLocalPlayback(options));
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void destroy()
         {
-            if(manager != null)
-            {
-                manager.Dispose();
-                manager = null;
-            }
+            manager?.SwitchPlayback(null);
         }
 
         [ReactMethod]
         public void updateOptions(JObject options, IPromise promise)
         {
-            // TODO remove the necessity of setupPlayer
-            manager?.UpdateOptions(options);
+            manager.UpdateOptions(options);
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void play(IPromise promise)
         {
-            manager?.GetPlayer()?.Play();
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.Play();
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void pause(IPromise promise)
         {
-            manager?.GetPlayer()?.Pause();
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.Pause();
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void stop(IPromise promise)
         {
-            manager?.GetPlayer()?.Stop();
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.Stop();
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void reset(IPromise promise)
         {
-            manager?.GetPlayer()?.Reset();
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.Reset();
+            promise.Resolve(null);
+        }
+
+        [ReactMethod]
+        public void removeUpcomingTracks(IPromise promise)
+        {
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.RemoveUpcomingTracks();
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void add(JArray array, string insertBeforeId, IPromise promise)
         {
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
             List<Track> tracks = new List<Track>(array.Count);
 
             foreach (JObject obj in array)
@@ -122,12 +153,15 @@ namespace TrackPlayer
                 tracks.Add(new Track(obj));
             }
 
-            manager?.GetPlayer()?.Add(tracks, insertBeforeId, promise);
+            player.Add(tracks, insertBeforeId, promise);
         }
 
         [ReactMethod]
         public void remove(JArray array, IPromise promise)
         {
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
             List<string> tracks = new List<string>(array.Count);
 
             foreach (string id in array)
@@ -135,59 +169,75 @@ namespace TrackPlayer
                 tracks.Add(id);
             }
 
-            manager?.GetPlayer()?.Remove(tracks, promise);
+            player.Remove(tracks, promise);
         }
 
         [ReactMethod]
         public void skip(string track, IPromise promise)
         {
-            manager?.GetPlayer()?.Skip(track, promise);
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.Skip(track, promise);
         }
 
         [ReactMethod]
         public void skipToNext(IPromise promise)
         {
-            manager?.GetPlayer()?.SkipToNext(promise);
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.SkipToNext(promise);
         }
 
         [ReactMethod]
         public void skipToPrevious(IPromise promise)
         {
-            manager?.GetPlayer()?.SkipToPrevious(promise);
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.SkipToPrevious(promise);
+        }
+
+        [ReactMethod]
+        public void getQueue(IPromise promise)
+        {
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            var queue = player.GetQueue();
+            var array = new JArray();
+            
+            foreach(var track in queue)
+            {
+                array.Add(track.ToObject());
+            }
+            
+            promise.Resolve(array);
         }
 
         [ReactMethod]
         public void getCurrentTrack(IPromise promise)
         {
-            Playback player = manager?.GetPlayer();
+            var player = manager?.GetPlayer();
             if (Utils.CheckPlayback(player, promise)) return;
-
-            string id = player.GetCurrentTrack()?.id;
-
-            if (id != null)
-                promise.Resolve(id);
-            else
-                promise.Reject("track", "No track playing");
+            
+            promise.Resolve(player.GetCurrentTrack()?.Id);
         }
 
         [ReactMethod]
         public void getTrack(string id, IPromise promise)
         {
-            Playback player = manager?.GetPlayer();
+            var player = manager?.GetPlayer();
             if (Utils.CheckPlayback(player, promise)) return;
 
-            JObject track = player.GetTrack(id)?.ToObject();
-
-            if (track != null)
-                promise.Resolve(track);
-            else
-                promise.Resolve(null);
+            promise.Resolve(player.GetTrack(id)?.ToObject());
         }
 
         [ReactMethod]
         public void getVolume(IPromise promise)
         {
-            Playback player = manager?.GetPlayer();
+            var player = manager?.GetPlayer();
             if (Utils.CheckPlayback(player, promise)) return;
 
             promise.Resolve(player.GetVolume());
@@ -196,14 +246,17 @@ namespace TrackPlayer
         [ReactMethod]
         public void setVolume(double volume, IPromise promise)
         {
-            manager?.GetPlayer()?.SetVolume(volume);
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.SetVolume(volume);
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void getRate(IPromise promise)
         {
-            Playback player = manager?.GetPlayer();
+            var player = manager?.GetPlayer();
             if (Utils.CheckPlayback(player, promise)) return;
 
             promise.Resolve(player.GetRate());
@@ -212,21 +265,27 @@ namespace TrackPlayer
         [ReactMethod]
         public void setRate(double rate, IPromise promise)
         {
-            manager?.GetPlayer()?.SetRate(rate);
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.SetRate(rate);
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void seekTo(double seconds, IPromise promise)
         {
-            manager?.GetPlayer()?.SeekTo(seconds);
+            var player = manager?.GetPlayer();
+            if (Utils.CheckPlayback(player, promise)) return;
+
+            player.SeekTo(seconds);
             promise.Resolve(null);
         }
 
         [ReactMethod]
         public void getPosition(IPromise promise)
         {
-            Playback player = manager?.GetPlayer();
+            var player = manager?.GetPlayer();
             if (Utils.CheckPlayback(player, promise)) return;
 
             promise.Resolve(player.GetPosition());
@@ -235,7 +294,7 @@ namespace TrackPlayer
         [ReactMethod]
         public void getBufferedPosition(IPromise promise)
         {
-            Playback player = manager?.GetPlayer();
+            var player = manager?.GetPlayer();
             if (Utils.CheckPlayback(player, promise)) return;
 
             promise.Resolve(player.GetBufferedPosition());
@@ -244,7 +303,7 @@ namespace TrackPlayer
         [ReactMethod]
         public void getDuration(IPromise promise)
         {
-            Playback player = manager?.GetPlayer();
+            var player = manager?.GetPlayer();
             if (Utils.CheckPlayback(player, promise)) return;
 
             promise.Resolve(player.GetDuration());
@@ -253,7 +312,7 @@ namespace TrackPlayer
         [ReactMethod]
         public void getState(IPromise promise)
         {
-            Playback player = manager?.GetPlayer();
+            var player = manager?.GetPlayer();
             if (Utils.CheckPlayback(player, promise)) return;
 
             promise.Resolve(player.GetState());
